Move config.py and log.py into processor directory
