+++ families/battleship/sawtooth_battleship/processor/main.py
@@ -19,10 +19,10 @@ import argparse
 import pkg_resources
 
 from sawtooth_sdk.processor.core import TransactionProcessor
-from sawtooth_sdk.client.config import get_log_dir
-from sawtooth_sdk.client.config import get_log_config
-from sawtooth_sdk.client.log import init_console_logging
-from sawtooth_sdk.client.log import log_configuration
+from sawtooth_sdk.processor.config import get_log_dir
+from sawtooth_sdk.processor.config import get_log_config
+from sawtooth_sdk.processor.log import init_console_logging
+from sawtooth_sdk.processor.log import log_configuration
 from sawtooth_battleship.processor.handler import BattleshipTransactionHandler
