+++ sdk/examples/intkey_python/sawtooth_intkey/processor/main.py
@@ -18,10 +18,10 @@ import argparse
 import pkg_resources
 
 from sawtooth_sdk.processor.core import TransactionProcessor
-from sawtooth_sdk.client.log import init_console_logging
-from sawtooth_sdk.client.log import log_configuration
-from sawtooth_sdk.client.config import get_log_config
-from sawtooth_sdk.client.config import get_log_dir
+from sawtooth_sdk.processor.log import init_console_logging
+from sawtooth_sdk.processor.log import log_configuration
+from sawtooth_sdk.processor.config import get_log_config
+from sawtooth_sdk.processor.config import get_log_dir
 from sawtooth_intkey.processor.handler import IntkeyTransactionHandler
