+++ sdk/python/sawtooth_sdk/client/__init__.py
@@ -12,9 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ------------------------------------------------------------------------------
-
-
-__all__ = [
-    'config',
-    'log'
-]
