+++ sdk/python/sawtooth_sdk/processor/log.py
